Smoke tests v2v
