+++ cfme/tests/v2v/test_v2v_migrations.py
@@ -18,7 +18,6 @@ from cfme.fixtures.v2v_fixtures import get_migrated_vm
 from cfme.infrastructure.provider.rhevm import RHEVMProvider
 from cfme.infrastructure.provider.virtualcenter import VMwareProvider
 from cfme.markers.env_markers.provider import ONE_PER_VERSION
-from cfme.utils.blockers import BZ
 
 pytestmark = [
     test_requirements.v2v,
@@ -42,7 +41,6 @@ pytestmark = [
 @pytest.mark.parametrize(
     "mapping_data_vm_obj_single_datastore",
     [
-        ["nfs", "nfs", rhel7_minimal],
         ["nfs", "iscsi", rhel7_minimal],
         ["iscsi", "iscsi", rhel7_minimal],
         ["iscsi", "nfs", rhel7_minimal],
@@ -327,50 +325,3 @@ def test_migrations_different_os_templates(
     for src_vm in src_vms_list:
         migrated_vm = get_migrated_vm(src_vm, provider)
         soft_assert(src_vm.mac_address == migrated_vm.mac_address)
-
-
-@pytest.mark.parametrize(
-    "v2v_provider_setup, mapping_data_multiple_vm_obj_single_datastore",
-    [["SSH", ["nfs", "nfs", [rhel7_minimal]]]],
-    indirect=True,
-)
-@pytest.mark.meta(blockers=[BZ(1707983, forced_streams=['5.10'])])
-def test_single_vm_migration_with_ssh(
-    request, appliance, provider, mapping_data_multiple_vm_obj_single_datastore
-):
-    """
-    Polarion:
-        assignee: sshveta
-        caseimportance: high
-        caseposneg: positive
-        testtype: functional
-        startsin: 5.10
-        casecomponent: V2V
-        initialEstimate: 1h
-    """
-    infrastructure_mapping_collection = appliance.collections.v2v_infra_mappings
-    mapping_data = mapping_data_multiple_vm_obj_single_datastore.infra_mapping_data
-    mapping = infrastructure_mapping_collection.create(**mapping_data)
-
-    @request.addfinalizer
-    def _cleanup():
-        infrastructure_mapping_collection.delete(mapping)
-
-    migration_plan_collection = appliance.collections.v2v_migration_plans
-    migration_plan = migration_plan_collection.create(
-        name="plan_{}".format(fauxfactory.gen_alphanumeric()),
-        description="desc_{}".format(fauxfactory.gen_alphanumeric()),
-        infra_map=mapping.name,
-        target_provider=provider,
-        vm_list=mapping_data_multiple_vm_obj_single_datastore.vm_list,
-    )
-
-    assert migration_plan.wait_for_state("Started")
-    assert migration_plan.wait_for_state("In_Progress")
-    assert migration_plan.wait_for_state("Completed")
-    assert migration_plan.wait_for_state("Successful")
-
-    # validate MAC address matches between source and target VMs
-    src_vm = mapping_data_multiple_vm_obj_single_datastore.vm_list.pop()
-    migrated_vm = get_migrated_vm(src_vm, provider)
-    assert src_vm.mac_address == migrated_vm.mac_address
